// frontend/src/lab/bindings/messaging_domain/declarations/messaging_domain.did.d.ts
/* eslint-disable */

// @ts-nocheck

// This file was automatically generated by @icp-sdk/bindgen@0.3.0.
// You should NOT make any changes in this file as it will be overwritten.
// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.

import type { ActorMethod } from '@icp-sdk/core/agent';
import type { IDL } from '@icp-sdk/core/candid';
import type { Principal } from '@icp-sdk/core/principal';

export interface Conversation {
  'id' : string,
  'participants' : Array<Principal>,
  'team_id' : [] | [string],
  'next_sequence' : bigint,
  'club_id' : string,
}
export interface Init { 'governor' : Principal }
export interface Message {
  'id' : string,
  'conversation_id' : string,
  'body' : string,
  'sender' : Principal,
  'sequence' : bigint,
  'idempotency_key' : string,
}
export interface MessagePage {
  'messages' : Array<Message>,
  'latest_sequence' : bigint,
  'next_sequence' : [] | [bigint],
}
export interface Receipt {
  'conversation_id' : string,
  'read' : boolean,
  'user' : Principal,
  'message_id' : string,
}
export type Result = { 'Ok' : Conversation } |
  { 'Err' : string };
export type Result_1 = { 'Ok' : Message } |
  { 'Err' : string };
export type Result_2 = { 'Ok' : Receipt } |
  { 'Err' : string };
export type Result_3 = { 'Ok' : State } |
  { 'Err' : string };
export type Result_4 = { 'Ok' : MessagePage } |
  { 'Err' : string };
export type Result_5 = { 'Ok' : Unread } |
  { 'Err' : string };
export type Result_6 = { 'Ok' : Message } |
  { 'Err' : string };
export interface State {
  'messages' : Array<Message>,
  'schema' : number,
  'unread' : Array<Unread>,
  'governor' : Principal,
  'conversations' : Array<Conversation>,
  'receipts' : Array<Receipt>,
}
export interface Unread {
  'conversation_id' : string,
  'count' : bigint,
  'user' : Principal,
  'last_read_sequence' : bigint,
}
export interface _SERVICE {
  'create_conversation' : ActorMethod<
    [string, [] | [string], Array<Principal>],
    Result
  >,
  'delete_message' : ActorMethod<[string], Result_6>,
  'export_state' : ActorMethod<[], Result_3>,
  'list_messages' : ActorMethod<[string, [] | [bigint]], Array<Message>>,
  'list_messages_page' : ActorMethod<[string, [] | [bigint], number], Result_4>,
  'mark_read' : ActorMethod<[string, string], Result_2>,
  'send_message' : ActorMethod<[string, string, string], Result_1>,
  'unread_count' : ActorMethod<[string], Result_5>,
}
export declare const idlFactory: IDL.InterfaceFactory;
export declare const init: (args: { IDL: typeof IDL }) => IDL.Type[];